Refactor file names to keep consistency
